requirements.txt
mkdocs-material==7.3.6
fancyboxmd==1.0.1
PyGithub==1.55
python-dotenv==0.19.1
